[global] Initial commit of rig
This represents the initial commit of the rig project.

Rig is meant to be a lightweight, flexible utility for system
administrators and support engineers to allow for easier troublshooting
of ongoing issues that either arise at random or inconvenient times that
make manual data collection difficult or impossible.

It is designed to be easily extended to allow for numerous types of
events occuring on the system (such as logged messages or network
activity) and trigger pre-selected actions such as collecting diagnostic
information.

In other words, rig is meant to assist with troubleshooting situations
that amount to "wait for this behavior to occur again and then do XYZ".

It should be noted that this project is still very much in its infancy
should be considered super-pre-alpha software.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>