[rigs|actions] Add priority weighting to order of action execution
Actions now have a priority attribute that will be used as a weighting
system to determine the order of execution when multiple actions are
specified by the user.

By default, an action's priority is set to 10. If a particular action
should execute _before_ any actions that use the default (i.e. they are
time-sensitive and need to be executed as close as possible to the time
of the triggering event), the value should be _lowered_. If the action
should be executed after other, more time-sensitive, actions have been
executed, then this value should be raised.

The noop action will use this default value and can be used as a good
benchmark for when yet-to-be-created actions should be prioritized. The
sosreport action will be given a starting priority of 100 - mostly
because the amount of time sosreport can take to run could present
issues if it were executed before other time-sensitive actions.

Also included in this is a minor fixup for logging report_results().

Resolves: #1

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>