[gcore] Add new action
Adds a new action that captures a coredump via GDB's gcore utility for a
process or processes.

This action is enabled by the --gcore option, which takes either a PID
directly or a process name which will then be converted to a PID. In the
event that a process name returns multiple PIDs, the default behavior is
to fail rig creation as we cannot determine which PID is the proper one
to use. This can be overridden with the --all-pids action option, in
which case gcore will be run against all returned PIDs for a process
name. The --gcore option can be specified multiple times.

This action has a priority weight of 1 - meaning it should run prior to
any other actions specified by the user.

Additionally, it should be noted that gcore does _not_ interrupt a
running process so this should not cause any impact in a production
environment.

Resolves: #2

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>