[rigs] Don't trigger actions if watcher thread returns False
When a watcher thread returns True, this is intended to mean the trigger
condition is met, and thus the specified actions should be taken. The
corollary of this is that if a watcher thread returns False, the trigger
condition has not been met and we should take no action. The implication
here is that something caused the rig to stop monitoring a resource.

For now, we'll continue with this mindset and this patch prevents a
watcher thread that returns False from triggering actions. This in turn
means that a watcher thread that does this will cause the entire rig to
be destroyed as we act on the first watcher thread to complete.

We'll need to keep this in mind going forward and see if we need to
adjust this approach.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>