[tcpdump] Add new action
Adds a new action that begins a tcpdump packet capture when the rig is
started, and stops the capture when the rig is triggered.

Users can provide any tcpdump-accepted filter string via the --filter
option to reduce the amount of network traffic captured.

This action will trigger before most other actions as it has a priority
of 2, which means currently only the gcore action will trigger before
it.

Resolves: #4

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>