[system] Add new rig
Adds a new 'system' rig that will monitor system resource consumption,
much like how the process rig monitors consumption for processes.

The system rig can watch CPU statistics such as iowait, steal, nice,
guest, and user as well as memory statistics such as total memory
available, used, free, or in slab.

With the exception of the available and free memory stats, each of the
above will trigger if the utilization rate is _above_ the defined
threshold for the rig. For available and free memory, the rig is
triggered if the system is _below_ the defined threshold.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>