[gcore] Update action to use get_option() method
Updates the gcore action to use the new get_option() method and to also
use the process rig's watched pid(s) in the event the '--gcore' option
is not given an explicit value.

Also improved the function to convert commandline values to pids and
checks to make sure the pid(s) exist.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>