[logs] Perform regex checks during rig initialization
Adds some basic checks to the provided 'message' option for the logs
rig. These checks are to ensure regex is syntactically correct and
parseable by the re module. The setup processes will now also check to
make sure the provided string will at least trigger against itself
before launching the rig. In either case if there is an issue, rig
creation aborts.

Further, this check highlights the need to use python-style regex
instead of shell-syntax. For the moment, rig will automatically convert
'*' usage to '.*' as that should be a harmless automatic change and
would be the most used shell-style regex syntax in log message matching.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>