[rig] Don't fail on list when /var/run/rig does not exist
Fixes an issue where a traceback would be printed to console if 'rig
list' was run when /var/run/rig did not exist (e.g. before any rigs are
created, ever). Instead, just set the socket list to empty and proceed
normally as if no rigs exist but the directory does (which will then be
created on the first rig creation).

Fixes: #17

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>