[ping] Add new rig
This adds a new rig type that performs an ongoing ping test against a
specified host. When a defined threshold of packets are lost or
exceeding a defined maximum RTT, the rig is triggered. Most errors
reported from the ping binary are considered fatal errors to the rig and
do NOT cause it to trigger.

This new rig violates the standing, though unwritten, policy of not
shelling-out to system binaries for rigs (meaning only actions may
shell-out generally), and it will be closely monitored for stability and
supportability. Generally speaking the rig project wants to keep
monitors as pure python for ease of ongoing maintenance and thread
handling. However, given that this is an extremely common use case, and
that we are sending a single individual ping at a time (rather than
monitoring a continuous uninterrupted call to ping), it should be a
relatively safe exception to make for something that will likely be a
highly used rig type.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>