[rigs|actions] Optionally repeat configured actions
This patch adds the ability to have certain actions be repeated during
the execution of a rig's configured actions. By using the --repeat
option, users may now specify a number of times actions should be
repeated. Using --repeat-delay will cause rig to pause for the given
amount of seconds before repeating an executed action.

At the time of this change, only the gcore action is repeatable.

Related: #13

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>