[build] Cut first beta version
This marks the first beta version for rig, following several months of
alpha development.

At this point, there are several types of monitors and actions that
cover a wide variety of use cases in a support or system administration
context.

This beta version's main focus will be suitability testing and bug
identification. There is not a set timeline at the moment for a
specific number of beta releases, release candidates, or set in stone
timeline before a GA release as the need for an as stable as possible
utility is paramount.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>