[rigs] Allow manually triggering rigs from the cmdline
Adds a new 'trigger' sub-command that allows a user to manually trigger
a specific rig.

The syntax for this is 'rig trigger -i $rig_id'. This will send a
request to the rig's listening socket which will then cause a new
default watcher thread to immediately trigger execution of the rig.

This sub-command does not accept a value of 'all', each rig must be
individually triggered.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>