[rigs] Change order of priority of get_option() return
If an option has already been called by get_option(), it would appear in
the rig_options dict. Until now, that dict was only checked if the
option was not specified on the cmdline - this in itself is an incorrect
flow, but this also prevented set_option() from actually being
effective.

Now, if the option appears in the rig_options dict, it is returned
before defaulting to the cmdline setting.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>