[BaseRig] Import last element of module inspect
The inspect done on discovered modules in _load_supported_actions() will
have any of our defined exceptions listed as well if the action imports
those exceptions.

Update the index that we use to identify the actual action class to be
the last list element to properly use the element to initialize the
action in question.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>