[sosreport] Add sanity check to --sos-opts
Adds a sanity check to the sosreport action via pre_action() that
inspects any passed value to --sos-opts for potential shell-code. This
is an effort to avoid chaining potentially malicious actions off of the
sosreport command invoked by rig. While this concern is admittedly minor
given that both rig and sosreport require root to function to begin
with, it is still something we should address.

If any shell-code is thought to be in the passed --sos-opts value, rig
creation is aborted. If none is found, we run the option through
pipes.quote() to add another layer of protection against undesired
behavior.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>