[sosreport] Replace --sos-opts with specific passthru options
Using --sos-opts presented a number of issues with regards to
potentially damaging behavior with specifically crafted strings, that
would be a cat-and-mouse game of preventing going forward.

Now, we expose a select number of sosreport options - namely the plugin
enablement options and --plugin-option - and pass that through the base
sanity check and finally quote the value to mitigate these kinds of
behaviors.

For non-exposed options, users should look to using /etc/sos.conf to
modify the default behavior of sosreport on the system rig is generating
an sosreport from.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>