[filesystem] Add new rig
Adds a new filesystem rig that will watch a specified filesystem,
directory, or file for size related triggers.

The most obvious use of this rig is to monitor a specific file, such as
a log file, for growing to a certain size - however the filesystem
backing whereever that file is may also be monitored in the same manner.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>