[rigs|actions] Add global --interval option
Adds a new --interval global option that can be used to specify the
amount of time, in seconds, to wait between rig polling operations.

Updates the existing rigs to use a new wait_loop() helper that utilizes
this new option, removing the need for individual rigs to deal with the
option directly.

Note that for the ping rig, this removes the --ping-interval option.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>