[rig] Move action parser setup to classmethod, add --name option
First, move the addition of Action parser options to classmethods so
that we don't need to initialize eveyr action everytime we start a rig,
actions are now only initialized if they are requested by the user.

Second, add a `--name` option that allows a manually specified rig ID,
rather than only having a randomly generated one.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>