[rig] Move loading of rigs out of Rigging()
This moves the loading of supported rigs out of the `Rigging()` class
and into the main 'binary'. This allows us to dynamically update the
parser. One drawback is that we will attempt to load the supported rig
modules even if we are just querying running rigs, however this
alleviates the burden of manually maintaining the parser usage output
for each newly added rig type.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>