[rig] Remove temp directory when `destroy` is called
Manually destroying a rig was not removing the temp directory created
during rig creation. This was because of a too-strict check for a
generated archive name, which is generally needed to prevent the
accidental deletion of data if for some reason we fail to compress data
present in the temp directory.

Now, properly remove the temp directory when `destroy` is called.

Closes: #26

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>