[rig] Fix status retrieval for foreground rigs
Previously, foreground rigs would not return usable data when calling
for the rig's status. This was due to the fact that we were only getting
the rig's pid when we detached from console.

Now, get the rig's pid during initialization so that foreground rigs can
return the proper status information.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>