[logs] Do not attempt to convert shell-style globs
The previous attempts to convert '*' to '.*' when taking regex strings
for the logs rig trigger message causes more frustration than it solves,
when a user passes a valid python-regex string.

Remove the attempt to convert, and rely on the existing messaging around
needing to use pythonic regex for this option.

Resolves: #29

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>