[tcpdump] Make snaplength configurable
The snaplength parameter for tcpdump is now user-controllable.
Previously, it was hardset to a value of '0' ensuring the entirety of
every packet was captured, but this is not always needed.

Now, via the `--snaplen` or `--snapshot-length` option, users may
control this setting passed to the tcpdump process kicked off by rig.

Resolves: #33

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>