[process] Wait for all monitored PIDs to exit
When watching multiple PIDs, a rig would trigger on the first PID to
exit if the rig was watching for anything other than process state of
not running. Meaning we could inadvertantly trigger the rig when we
should have instead continued monitoring the remaining PIDs being
watched.

Fix that, by blocking the threads that would otherwise exit due to a
killed PID.

Resolves: #36

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>