[packet] Add new `packet` rig + documentation
This rig is a packet sniffer. It will trigger whenever a network
packet matching the provided attributes is seen on any of the system
interfaces. `--ifname` can be specified to ignore packets seen on
interfaces other than the one defined.

The rig doesn't put the interfaces in promiscuous mode. Therefore only
traffic where the source or destination is the host itself, or traffic with
a multicast or broadcast destination will be seen.

By default, all the provided attributes must match for the rig to trigger.
If `--any` is specified, then matching only one of the provided attributes
is enough.

`--payload` is provided for convenience to match application payload on
cleartext protocols. For example, it can be used to match HTTP requests or
status codes, or DNS requests. However doing payload matching on systems
with a high network traffic load will be resource intensive.