[rig] Use `tempfile` module for temp directory creation
Previously, a change was made to temp directory creation in an effort to
make it more secure. While that was largely handled, it left us with an
unhandled error in an edge case configuration. Rather than putting a
band-aid over that again, re-write the temp directory creation process
to leverage the `tempfile` module, so that we can safely and completely
ignore the id/name of a rig, and leave the use of that for the
communication socket.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>