[rig] Cleanup tempdir even if no data generated
Previously, we would erroneously not remove a rig's temp dir if it
generated no data. Specifically, this could mean left over temp
directories when using the `--noop` action to test rig configurations.

Fix this by not relying on an archive name to remove the temp dir, and
instead only leave the directory in place if we are specifically not
building an archive.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>