[rig] Set initial framework for new command structure
This commit lays the groundwork for a new approach to rig CLI commands,
e.g. a user running `rig info`.

This new approach abstracts the handling of command strings compared to
the current method, giving each command its own subparser for which to
control its options. In this way, we will no longer need to account for
colliding option names.

Note that this is a stepping stone commit, and rig is not considered
functional at this specific point in time as we transition to the new
2.0 design. Future commits will rectify this.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>