[process] Port `process` rig to new monitor design
Ports the previous `process` rig to the new monitor design for rig v2.
The new monitor accepts either a single process or a list of processes
specified by either PID(s) or process/command name(s). Process names
can be exact strings or python-compatible regexes.

One difference between the new monitor and the old rig, is that the
`--all` option has not been brought forward to the new monitor. It is
assumed that any process names given to this monitor is done to
intentionally catch all processes matching that name.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>