[rigs] Drop remaining `rigging.rigs` package from v1
Both `BaseRig` and the existing rigs from the previous v1 designs have
been transitioned to the new v2 design of rig, whereby a "rig" now
consitutes a collection of monitor(s) and action(s), and the previous
rigs have been ported to monitors.

Note that the `ping` rig is not being moved forward at this time, as it
is better replaced by the `packet` rig.

Drop the lingering bits in `rigging.rigs` from the tree due to this part
of the transition being complete.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>