[connection] Add a new class for communicating with rigs
This commit adds an abstraction that can be used by rig subcommands in
order to communicate with an existing rig over the socket used by that
rig.

The new `RigConnection` class will do basic checks for a rig of a given
name, and will handle formatting the request and the response to/from
the rig's socket via a `multiprocessing.connection` Client.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>