[destroy] Re-implement the `destroy` command
Re-implements the `destroy` command for the new design so that users can
run `rig destroy` and terminate a rig without triggering its actions or
generating a rig archive.

This is the first command to make sure of socket communication with
rigs, and as such makes changes to `BaseRig` such as flipping us over to
using a `multiprocessing.connection` Listener rather than directly
leveraging and manipulating the socket ourselves.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>