[rigging] Fix potential scenario where Listener deletes socket on fork
When forking to background the rig process, the Listener may release the
socket, thus causing it to be removed from the filesystem. This rig
continues to run, but in such a situation we no longer have a way to
communicate with it.

Refactor the initial socket check and how/when we create the Listener
socket so that we can still avoid duplicating rigs by name, and at the
same time ensure the socket exists while the backgrounded rig process is
running.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>