[timer] Add new timer monitor
This commit adds a new `timer` monitor, which simply serves to provide a
maximum run time for the rig. This monitor accepts two parameters in a
rigfile:

  1. `timeout`, which specifies how much time to wait until the monitor
triggers. This parameter can take either a raw value in seconds, or a
more human-friendly string such as `1h 30m` for example.

  2. `trigger_on_expiry`, which controls if the timeout expiring
triggers the rig, or simply causes the rig to be removed without
triggering actions. Default is true (trigger actions on timeout).

Resolves: #45

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>