[build] Update build scripts for new design version
This commit updates the build script and spec file for the 2.0 version
redesign. It is not intended to be built and distributed in downstream
yet, so it is marked as '2.0a' to indicate it is in an alpha state. This
is being done now to allow local builds to function and consume the
recently added documentation for the new sub commands.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>